refactor: :fire: added gitignore
