+++ .gitignore
@@ -1,0 +1,4 @@
+__pycache__
+src/__pycache__
+.vscode
+.env
